FundRatingNSDL/nsdl_extraction/requirements.txt
Scrapy==2.3.0
itemadapter==0.1.0
selenium==3.141.0
python_dateutil==2.8.1
